chore(Flow): split state viewer helpers into domain modules [#1752]
